Move previewArea.js
