+++ app/js/ui/previewArea.js
